v3

+++ src/main/java/iset/pfe/mediconnectback/repositories/DossierMedicalRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import iset.pfe.mediconnectback.entities.DossierMedical;
 
 @Repository
-public interface DossierMedicalRepository extends JpaRepository<DossierMedical, Integer> {
+public interface DossierMedicalRepository extends JpaRepository<DossierMedical, Long> {
     @Query("SELECT dm FROM DossierMedical dm JOIN FETCH dm.patient WHERE dm.patient.id = :patientId")
     List<DossierMedical> findByPatientId(@Param("patientId") Long patientId);
     @Query("SELECT dm FROM DossierMedical dm JOIN FETCH dm.patient p JOIN FETCH dm.fichiers f JOIN FETCH dm.medecin m WHERE dm.medecin.id = :medecinId")
